Monitoring Full Propagation of Topic in Cluster -- Cluster Topic Manipulation Service #271
1 - Monitoring Full Propagation of Topics within a Cluster -- Cluster Topic Manipulation Service. This service periodically creates and deletes topic and measures the duration from operation start to full propagation of the metadata to all the brokers.

All Kafka brokers can answer a topic metadata request that describes the current state of the cluster, such as what topics there are, which partitions those topics have, which broker is the leader for those partitions, and the host and port information for these brokers.
The client finds a broker and that broker tells the client regarding all the other brokers which exist and what partitions they host. (This first broker may itself go down so the best practice for a client is to take ask two or three bootstrap brokers).
The topic creation could be issued through AdminClient and we could have a scenario where we send a request to broker A for metadata fetch. The broker could fetch the metadata and return a good signal and but in reality the broker B could be experiencing extremely slow speed/network and so it might not have even finished creating these topic partitions (propagating metadata).
At the point in time, checking the log directories of broker B will return negative (bad) since the topic partitions haven’t been created for that partition.

In other words, cluster topic metadata is eventually consistent. when metadata updates are propagated by the controller, the brokers can see them in any order and there can be delays. the client deals with this by getting NO_TOPIC_OR_PARTITION or other exception messages, which triggers it to do the lookup cycle over again, re-attempting until it gets a response from the broker that is supposed to be the leader.

If we send broker a fetch topic metadata request -> let’s say A says that the topic exists in the metadata.
it can be the case that it turns out broker B is extremely slow due to bad network or heavy queue and it doesn’t have the topic partitions populated in its log directory.

2 - TODO: Metrics/Sensors to record the time it takes to finish this metadata change (topic creation). -- separate PR
3 - TODO: Metrics/Sensors to record the time it takes to finish this metadata change (topic deletion). -- separate PR
4 - TODO in a sep. PR: make runnable interval configurable.
Andrew Choi