Fix replicasAssignment Hashmap (#340)
Fix for the issue #325. 
When creating a ClusterTopicManipulationService topic, the constructor uses a Hashmap containing the brokers id for each replica partition.
The problem is that the same broker id can be repeated.
With this fix, a simple shuffle of the brokers allows a random repartition of the brokers ids.
No need for the while loop.